moved fragmets inside Fragment folder
