+++ app/src/main/java/com/example/poseexercise/views/fragment/HomeFragment.kt
@@ -1,10 +1,11 @@
-package com.example.poseexercise
+package com.example.poseexercise.views.fragment
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.poseexercise.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
