+++ app/src/main/res/layout/fragment_bottom_navigation.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BottomNavigationFragment">
+    tools:context=".views.fragment.BottomNavigationFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
