+++ app/src/main/res/layout/fragment_home.xml
@@ -4,8 +4,8 @@
     android:id="@+id/homeFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/primaryColor"
-    tools:context=".HomeFragment">
+    android:background="@color/white"
+    tools:context=".views.fragment.HomeFragment">
 
     <TextView
         android:layout_width="match_parent"
