+++ app/src/main/res/layout/fragment_workout.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#FFC0FF"
-    tools:context=".WorkoutFragment">
+    tools:context=".views.fragment.WorkoutFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
